requirements.txt
python-dotenv==1.0.1
Flask==3.0.3
soundfile==0.12.1
TTS==0.22.0